feat(ini-005): catalogue-tooling Wave 5a — repository rewire (Makefile + tools/ reorganisation)
Rewires the home repo to use the new canonical agentbundle catalogue *
commands and reorganises tools/ into tools/catalogue/ + tools/repo/.

Engine-Change-RFC: ini-005
