+++ .github/workflows/docs.yml
@@ -111,8 +111,8 @@ jobs:
   # Aggregator end-to-end. Duplicates the per-layer linter jobs by design —
   # those each surface with their own check name in PR UI; this one catches
   # regressions in the catalogue gate's composition (a layer dropped, a reorder
-  # breaking a dependency, exit-code propagation). Runs the repo-native
-  # `tools/pre-pr-catalogue.py` (the 8 catalogue checks + the shipped pre-pr.py).
+  # breaking a dependency, exit-code propagation). Runs agentbundle catalogue
+  # verify (portable) then the repo-specific catalogue checks + shipped pre-pr.py.
   hooks:
     name: Lifecycle hooks
     runs-on: ubuntu-latest
@@ -126,7 +126,7 @@ jobs:
           sed 's/"pending"/"approved"/' .claude/skills/work-loop/assets/state.json \
             > docs/specs/_ci-smoke/state.json
       - name: Run pre-pr aggregator (mirrors local enforcement triplet)
-        run: python tools/pre-pr-catalogue.py
+        run: PYTHONPATH=packages/agentbundle python tools/catalogue/pre_pr_catalogue.py
       - name: Self-test the aggregator (each layer corruption surfaces)
         run: bash tools/test-pre-pr.sh
       - name: Self-test session-start (scope filter + malformed warning)
